Change `size_t` to `INT` for index types in `interpolation_data.hpp` and `interp_utils.hpp` as well as for any downstream mock data/tests.

+++ src/reactions_lib/interp_utils.hpp
@@ -9,7 +9,7 @@ namespace VANTAGE::Reactions::interp_utils {
  * Helper macro that extracts the value of the binary representation of i at
  * position j (in the binary representation of i).
  */
-template <typename T> inline T binary_extract(const T &i, const size_t &j) {
+template <typename T> inline T binary_extract(const T &i, const INT &j) {
   return ((i >> j) & 1);
 }
 
@@ -22,15 +22,14 @@ template <typename T> inline T binary_extract(const T &i, const size_t &j) {
  * @param dims_vec Pointer to a vector that contains the size of each dimension.
  * @param ndim The number of dimensions
  *
- * @return size_t that specifies the index on a contiguous grid array
+ * @return INT that specifies the index on a contiguous grid array
  */
-inline size_t coeff_index_on_device(size_t const *indices,
-                                    size_t const *dims_vec,
-                                    const size_t &ndim) {
-  size_t index = indices[ndim - 1];
+inline INT coeff_index_on_device(INT const *indices, INT const *dims_vec,
+                                 const INT &ndim) {
+  INT index = indices[ndim - 1];
 
   // slight re-factor to avoid underflow errors with size_t.
-  size_t dimx = ndim - 1;
+  INT dimx = ndim - 1;
   while (dimx > 0) {
     --dimx;
     index *= dims_vec[dimx];
@@ -50,14 +49,13 @@ inline size_t coeff_index_on_device(size_t const *indices,
  * dimension of a 4D grid, the dim_index=2
  * @param dims_vec Pointer to a vector that contains the size of each dimension.
  *
- * @return size_t that specifies the index on a contiguous ranges array.
+ * @return INT that specifies the index on a contiguous ranges array.
  */
-inline size_t range_index_on_device(const size_t &sub_index,
-                                    const size_t &dim_index,
-                                    size_t const *dims_vec) {
-  size_t index = sub_index;
+inline INT range_index_on_device(const INT &sub_index, const INT &dim_index,
+                                 INT const *dims_vec) {
+  INT index = sub_index;
 
-  for (size_t i = 0; i < dim_index; i++) {
+  for (INT i = 0; i < dim_index; i++) {
     index += dims_vec[i];
   }
 
@@ -74,15 +72,14 @@ inline size_t range_index_on_device(const size_t &sub_index,
  * given dimension.
  * @param last_index The last index in the range of the given dimension.
  *
- * @return size_t The index on a given dimension that is the closest to
+ * @return INT The index on a given dimension that is the closest to
  * x_interp.
  */
-inline size_t calc_floor_point_index(const REAL &x_interp,
-                                     REAL const *dim_range,
-                                     const size_t &last_index) {
-  size_t L = 0;
-  size_t R = last_index;
-  size_t m;
+inline INT calc_floor_point_index(const REAL &x_interp, REAL const *dim_range,
+                                  const INT &last_index) {
+  INT L = 0;
+  INT R = last_index;
+  INT m;
 
   while ((R - L) > 1) {
     m = L + ((R - L) / 2);
@@ -136,11 +133,11 @@ inline REAL linear_interp(const REAL x_interp, const REAL x0, const REAL x1,
  * @return std::vector<INT> That contains the points denoting the vertices of
  * the hypercube.
  */
-inline std::vector<size_t> construct_initial_hypercube(const size_t &ndim) {
-  size_t total_num = 1 << ndim;
-  std::vector<size_t> points(total_num);
+inline std::vector<INT> construct_initial_hypercube(const INT &ndim) {
+  INT total_num = 1 << ndim;
+  std::vector<INT> points(total_num);
 
-  for (size_t i = 0; i < total_num; i++) {
+  for (INT i = 0; i < total_num; i++) {
     points[i] = (i ^ (i >> 1));
   }
 
@@ -169,11 +166,11 @@ inline std::vector<size_t> construct_initial_hypercube(const size_t &ndim) {
  * representation
  */
 inline void initial_func_eval_on_device(
-    REAL *vertex_func_evals, size_t *vertex_coord, REAL const *func_grid,
-    size_t const *hypercube_vertices, size_t const *origin_indices,
-    size_t const *dims_vec, const size_t &ndim, const size_t &num_points) {
-  for (size_t point_index = 0; point_index < num_points; point_index++) {
-    for (size_t vertex_index = 0; vertex_index < ndim; vertex_index++) {
+    REAL *vertex_func_evals, INT *vertex_coord, REAL const *func_grid,
+    INT const *hypercube_vertices, INT const *origin_indices,
+    INT const *dims_vec, const INT &ndim, const INT &num_points) {
+  for (INT point_index = 0; point_index < num_points; point_index++) {
+    for (INT vertex_index = 0; vertex_index < ndim; vertex_index++) {
       vertex_coord[vertex_index] =
           origin_indices[vertex_index] +
           binary_extract(hypercube_vertices[point_index], vertex_index);
@@ -214,17 +211,16 @@ inline void initial_func_eval_on_device(
  */
 
 inline void contract_hypercube_on_device(
-    const REAL *interp_points, const size_t &dim_index,
-    size_t const *hypercube_vertices, const size_t *origin_indices,
-    const REAL *vertex_func_evals, REAL const *ranges_vec,
-    size_t const *dims_vec, REAL *output_evals, size_t *varying_dim,
-    size_t *vertex_coord) {
-  size_t ndim = dim_index + 1;
-  size_t num_points = (1 << ndim);
-  size_t num_out_points = (1 << dim_index);
+    const REAL *interp_points, const INT &dim_index,
+    INT const *hypercube_vertices, const INT *origin_indices,
+    const REAL *vertex_func_evals, REAL const *ranges_vec, INT const *dims_vec,
+    REAL *output_evals, INT *varying_dim, INT *vertex_coord) {
+  INT ndim = dim_index + 1;
+  INT num_points = (1 << ndim);
+  INT num_out_points = (1 << dim_index);
 
-  for (size_t point_index = 0; point_index < num_points; point_index++) {
-    for (size_t eval_index = 0; eval_index < ndim; eval_index++) {
+  for (INT point_index = 0; point_index < num_points; point_index++) {
+    for (INT eval_index = 0; eval_index < ndim; eval_index++) {
       vertex_coord[eval_index] =
           origin_indices[eval_index] +
           binary_extract(hypercube_vertices[point_index], eval_index);
@@ -232,10 +228,10 @@ inline void contract_hypercube_on_device(
     varying_dim[point_index] = vertex_coord[dim_index];
   }
 
-  size_t vertex_0, vertex_1;
+  INT vertex_0, vertex_1;
   REAL range_val_0, range_val_1, eval_point_0, eval_point_1;
 
-  for (size_t i = 0; i < num_out_points; i++) {
+  for (INT i = 0; i < num_out_points; i++) {
     vertex_0 = varying_dim[i];
     vertex_1 = varying_dim[num_points - (i + 1)];
